pyproject.toml
[tool.poetry]
name = "backup-devops"
version = "0.2.0"
description = ""
authors = ["eric-therond <eric.therond.fr@gmail.com>"]
readme = "README.md"
packages = [{include = "backup_devops"}]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10"
azure-devops = "^7.1.0b4"
gitpython = "^3.1.41"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
